+++ components/backbone.wreqr/src/wreqr.handlers.js
@@ -8,8 +8,13 @@ Wreqr.Handlers = (function(Backbone, _){
   // Constructor
   // -----------
 
-  var Handlers = function(){
-    this._handlers = {};
+  var Handlers = function(options){
+    this.options = options;
+    this._wreqrHandlers = {};
+    
+    if (_.isFunction(this.initialize)){
+      this.initialize(options);
+    }
   };
 
   Handlers.extend = Backbone.Model.extend;
@@ -17,24 +22,45 @@ Wreqr.Handlers = (function(Backbone, _){
   // Instance Members
   // ----------------
 
-  _.extend(Handlers.prototype, {
+  _.extend(Handlers.prototype, Backbone.Events, {
+
+    // Add multiple handlers using an object literal configuration
+    setHandlers: function(handlers){
+      _.each(handlers, function(handler, name){
+        var context = null;
+
+        if (_.isObject(handler) && !_.isFunction(handler)){
+          context = handler.context;
+          handler = handler.callback;
+        }
+
+        this.setHandler(name, handler, context);
+      }, this);
+    },
 
     // Add a handler for the given name, with an
     // optional context to run the handler within
-    addHandler: function(name, handler, context){
+    setHandler: function(name, handler, context){
       var config = {
         callback: handler,
         context: context
       };
 
-      this._handlers[name] = config;
+      this._wreqrHandlers[name] = config;
+
+      this.trigger("handler:add", name, handler, context);
+    },
+
+    // Determine whether or not a handler is registered
+    hasHandler: function(name){
+      return !! this._wreqrHandlers[name];
     },
 
     // Get the currently registered handler for
     // the specified name. Throws an exception if
     // no handler is found.
     getHandler: function(name){
-      var config = this._handlers[name];
+      var config = this._wreqrHandlers[name];
 
       if (!config){
         throw new Error("Handler not found for '" + name + "'");
@@ -48,12 +74,12 @@ Wreqr.Handlers = (function(Backbone, _){
 
     // Remove a handler for the specified name
     removeHandler: function(name){
-      delete this._handlers[name];
+      delete this._wreqrHandlers[name];
     },
 
     // Remove all handlers from this registry
     removeAllHandlers: function(){
-      this._handlers = {};
+      this._wreqrHandlers = {};
     }
   });