+++ components/marionette/lib/core/amd/underscore-mustache.js
@@ -1,9 +1,0 @@
-define(['underscore'],
-function(_) {
-    // override default global templates to use MUSTACHE style {{ name }} instead of ERB
-    _.templateSettings = {
-        evaluate: /\{\[([\s\S]+?)\]\}/g,
-        interpolate: /\{\{(.+?)\}\}/g
-    };
-    return _;
-});